refactor: move setup-babel to private directory
